+++ src/lib/mrvFl/mrvCallbacks.cpp
@@ -2217,6 +2217,12 @@ namespace mrv
                 return;
             }
         }
+        // If no next annotation was found, look to the first annotation
+        if (!times.empty())
+        {
+            view->stop();
+            player->seek(times[0]);
+        }
     }
 
     void next_annotation_cb(Fl_Menu_*, ViewerUI* ui)
@@ -2237,6 +2243,12 @@ namespace mrv
                 player->seek(time);
                 return;
             }
+        }
+        // If no next annotation was found, look to the first annotation
+        if (!times.empty())
+        {
+            view->stop();
+            player->seek(times[0]);
         }
     }