+++ src/docs/HISTORY.md
@@ -3,6 +3,8 @@ v1.3.7
 
 - Fixed Ghost Previous and Ghost Next not being in sync on the Secondary 
   viewport.
+- Made stepping with Next Annotation and Previous Annotation loop once they 
+  reach the final and first annotation respectively.
   
 v1.3.6
 ======